+++ stream-investment/investment-core/src/main/java/com/backbase/stream/configuration/InvestmentServiceConfiguration.java
@@ -5,6 +5,7 @@ import com.backbase.investment.api.service.v1.AllocationsApi;
 import com.backbase.investment.api.service.v1.AssetUniverseApi;
 import com.backbase.investment.api.service.v1.AsyncBulkGroupsApi;
 import com.backbase.investment.api.service.v1.ClientApi;
+import com.backbase.investment.api.service.v1.CurrencyApi;
 import com.backbase.investment.api.service.v1.FinancialAdviceApi;
 import com.backbase.investment.api.service.v1.InvestmentApi;
 import com.backbase.investment.api.service.v1.InvestmentProductsApi;
@@ -19,11 +20,13 @@ import com.backbase.stream.investment.service.CustomIntegrationApiService;
 import com.backbase.stream.investment.service.InvestmentAssetPriceService;
 import com.backbase.stream.investment.service.InvestmentAssetUniverseService;
 import com.backbase.stream.investment.service.InvestmentClientService;
+import com.backbase.stream.investment.service.InvestmentCurrencyService;
 import com.backbase.stream.investment.service.InvestmentIntradayAssetPriceService;
 import com.backbase.stream.investment.service.InvestmentModelPortfolioService;
 import com.backbase.stream.investment.service.InvestmentPortfolioAllocationService;
 import com.backbase.stream.investment.service.InvestmentPortfolioService;
 import com.backbase.stream.investment.service.resttemplate.InvestmentRestAssetUniverseService;
+import com.backbase.stream.investment.service.resttemplate.InvestmentRestDocumentContentService;
 import com.backbase.stream.investment.service.resttemplate.InvestmentRestNewsContentService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
@@ -34,6 +37,7 @@ import org.springframework.context.annotation.Import;
 
 @Import({
     DbsApiClientsAutoConfiguration.class,
+    InvestmentClientConfig.class
 })
 @EnableConfigurationProperties({
     InvestmentIngestionConfigurationProperties.class
@@ -97,6 +101,11 @@ public class InvestmentServiceConfiguration {
     }
 
     @Bean
+    public InvestmentCurrencyService investmentCurrencyService(CurrencyApi currencyApi) {
+        return new InvestmentCurrencyService(currencyApi);
+    }
+
+    @Bean
     public InvestmentSaga investmentSaga(InvestmentClientService investmentClientService,
         InvestmentPortfolioService investmentPortfolioService,
         InvestmentModelPortfolioService investmentModelPortfolioService,
@@ -112,17 +121,20 @@ public class InvestmentServiceConfiguration {
         InvestmentAssetUniverseService investmentAssetUniverseService,
         InvestmentAssetPriceService investmentAssetPriceService,
         InvestmentIntradayAssetPriceService investmentIntradayAssetPriceService,
+        InvestmentCurrencyService investmentCurrencyService,
         AsyncTaskService asyncTaskService,
         InvestmentIngestionConfigurationProperties coreConfigurationProperties) {
         return new InvestmentAssetUniverseSaga(investmentAssetUniverseService, investmentAssetPriceService,
-            investmentIntradayAssetPriceService, asyncTaskService, coreConfigurationProperties);
+            investmentIntradayAssetPriceService, investmentCurrencyService, asyncTaskService,
+            coreConfigurationProperties);
     }
 
     @Bean
     public InvestmentContentSaga investmentContentSaga(
         InvestmentRestNewsContentService investmentRestNewsContentService,
+        InvestmentRestDocumentContentService investmentRestDocumentContentService,
         InvestmentIngestionConfigurationProperties coreConfigurationProperties) {
-        return new InvestmentContentSaga(investmentRestNewsContentService, coreConfigurationProperties);
+        return new InvestmentContentSaga(investmentRestNewsContentService, investmentRestDocumentContentService, coreConfigurationProperties);
     }
 
 }