+++ stream-investment/investment-core/src/main/java/com/backbase/stream/investment/InvestmentContentData.java
@@ -1,5 +1,7 @@
 package com.backbase.stream.investment;
 
+import com.backbase.stream.investment.model.ContentDocumentEntry;
+import com.backbase.stream.investment.model.ContentTag;
 import com.backbase.stream.investment.model.MarketNewsEntry;
 import java.util.List;
 import lombok.Builder;
@@ -11,6 +13,9 @@ import lombok.EqualsAndHashCode;
 @Builder
 public class InvestmentContentData {
 
+    private List<ContentTag> marketNewsTags;
     private List<MarketNewsEntry> marketNews;
+    private List<ContentTag> documentTags;
+    private List<ContentDocumentEntry> documents;
 
 }