+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 ## [9.6.0]
 ### Added
 - customer-data-platform (CDP) module for the customer profiles ingestion
+ - Implement for Document ingestion;
+ - Implement Document tags ingestion;
+ - add Content news tags upsert;
+ - add currencies upsert;
+
+### Changed
+ - use latest investment service api 1.3.0;
+ - move customization of specs into an investment package;
 
 ## [9.5.0]
 ### Added